requirements.txt
mcp>=0.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0 